Apply incrementalActionSensitivity scaling to continuous-action trim inputs.
Also scale by Time.deltaTime for trim and incremental-throttle inputs, as otherwise
the rate of change would depend on framerate.

This makes trim and incremental throttle usable with hat axes (and, I guess, more usable
across the whole range of a proper analog axis). Previously trim would immediately clip
to +/-1 if you tried to control it with a hat axis, which is no use.

Default sensitivity might need increasing now that everything's scaled by the delta-time.